requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pytesseract
python-multipart
Pillow
pdf2image



